+++ src/main/resources/com/github/jlangch/venice/repl.sh
@@ -40,7 +40,7 @@ while true; do
     -Djava.io.tmpdir=${REPL_HOME}/tmp \
     -cp "libs:libs/*" \
     com.github.jlangch.venice.Launcher \
-    -loadpath "scripts" \
+    -loadpath "" \
     -restartable \
     -colors